chore: move validate condition to ha folder
