+++ src/utils/conditional/conditional-base.ts
@@ -1,10 +1,13 @@
 import { LovelaceCard } from "custom-card-helpers";
 import { PropertyValues, ReactiveElement } from "lit";
 import { customElement, property } from "lit/decorators.js";
+import {
+    checkConditionsMet,
+    validateConditionalConfig,
+} from "../../ha/panels/lovelace/common/validate-condition";
 import { HomeAssistant } from "../../ha/types";
 import { ConditionalChipConfig, LovelaceChip } from "../lovelace/chip/types";
 import { ConditionalCardConfig } from "../lovelace/types";
-import { checkConditionsMet, validateConditionalConfig } from "./validate-condition";
 
 @customElement("mushroom-conditional-base")
 export class ConditionalBase extends ReactiveElement {
